Add last_fetched_favorite_id
After every time we make a request to twitter to get a user’s
favorites, we’ll save the id of the last favorited tweet in the
response to the db. This way, on a subsequent request to twitter, we
can pick up where we left off and not get any favorites that we already
have. The reason I’m doing it this way is to allow users to delete any
favorites they don’t want from the db but not delete it in twitter.
That way, if a user deletes a favorite from the db and later makes a
request to get more favorites from twitter, we don’t end up re-fetching
and saving that favorite.