FavoritesRunner
The FavoritesRunner is responsible for grabbing a user’s favorited
tweets from the twitter api and persisting them to the db. It keeps
track of the id of the last favorited tweet in the api response in
order to grab *new* favorited tweets in subsequent requests.

It also keeps track of whether the last favorited tweet in the response
is that last of the user’s favorites in order to not make unnecessary
subsequent requests.